chore: rename new notes components
