+++ src/renderer/src/App.svelte
@@ -5,14 +5,14 @@
   import Pomodoro from './pages/PomodoroPage.svelte';
   import Calendar from './pages/CalendarPage.svelte'
   import Planner from './pages/PlannerPage.svelte';
-  import newNotes from './pages/newNotes.svelte';
+  import Notes from './pages/NotesPage.svelte';
 
   const routes = {
     '/': Dashboard,
     '/calendar': Calendar,
     '/planner': Planner,
     '/pomodoro': Pomodoro,
-    '/newNotes':newNotes
+    '/notes':Notes,
   }
   
 </script>
@@ -23,7 +23,7 @@
     <button><a href="#/calendar">Calendar</a></button>
     <button><a href="#/planner">Week Planner</a></button>
     <button><a href="#/pomodoro">Pomodoro</a></button>
-    <button><a href="#/newNotes">Notes</a></button>
+    <button><a href="#/notes">Notes</a></button>
   </nav>
 
   <Router {routes} />
