+++ src/renderer/src/lib/notes/Card.svelte
