+++ src/renderer/src/lib/notes/modal.svelte
