+++ src/renderer/src/lib/notes/notesDetails.svelte
