+++ src/renderer/src/lib/notes/notesList.svelte
