+++ src/renderer/src/lib/notes/stores/NotesStore.ts
